Medi_scan/requirements.txt
Django==5.0.6
Pillow==10.4.0
pytesseract==0.3.13
opencv-python==4.10.0.84
rapidfuzz==3.9.6
python-dotenv==1.0.1
